Remove followedTopic ids from the api query
This change is to update followedTopics properly when users click an external follow button. x-topic-search executes the topicFollowed action (which is triggered in the consumer side) when that happens. x-topic-search gets only the topic id from the event, but it needs the whole topic data to render correct info. Changing from [Fetch suggestions except followedTopics] to [Fetch all suggestions and then pick followedTopic up by the ids] will make easier to update followedTopics with the whole topic data