Noddy redux state comparison before running `setState()`
As we're not using immutable (or similar), we can't triple-equals match the current state with the next state, as the object reference will have changed even ifs keys and values have not.

This puts a basic check in to ensure if the current state shallow-matches the next state, we skip over it.