Call audio.load() on LOAD_MEDIA action to trigger preloading
This should deal with the delaying in audio being played seen on iOS/Safari devices.

It seems to be that Safari doesn't honour the `audio.preload` attribute until `audio.load()` is called. In other browsers, setting `audio.src` is enough.

Current use case is that the audio player is only shown after a user has requested to play audio, so preloading is ok here. In the future we may need to make this an option.