Remove oShare
The oShare component provides styles and a little JavaScipt for
sharing.

This component is bastardising the share styles quite a lot, and the
JavaScript is quite minimal (only used for popping sharing out in it's
own window) so I think we can remove it and get the share icons
directly from the image service.

This will mean the orgiami team are not blocked from changes to oshare
by this weird usecase.