Use a prop to determine podcast launcher visibility
As part of a bug raised via editorial, we need to hide the podcast app
links on desktop because they don't do anything.
To do this I have used a combination of @media query to show at narrow
screens (ie when the website is viewed on a mobile) and an isApp prop to
always show it if you're viewing in the app wrapper.