Ensure x-podcast-launchers has node-sass dependency.
It's not clear why this has started failing but perhaps since these packages are built in parallel there was a race condition that we were not aware of.

Hopefully by adding node-sass as a dependency to this package, which matches x-gift-article and x-follow-button, we'll resolve this issue and master will not fail.