Mitigate interference from `fastclick` plugin
On FT-App `fastclick` is included to allow double clicks not to be
misinterpreted as "double click to zoom". This is causing radio buttons
on the Consent Form to require two clicks to fire their `change` event.

A `needsclick` class is provided to allow clickable elements to opt out
of this behaviour. This PR applies them to the labels that form the hit
area for the radio inputs.