Updates to use this component at the client side
There are two major updates to use this component at the client side:

We need to render a property named `data-x-live-blog-wrapper-id` in the main element. This is now used to identify the element at the client side for live updates.

Previously we used `data-x-dash-id` for this purpose, but that property is only rendered when using this component with server side rendering.

Add `actions` argument to `listenToLiveBlogEvents` function. With server side rendering and hydrating at the client side, component actions can be triggered by dispatching a custom event named `x-interaction.trigger-actions`. This method can't be used with client side rendering. Therefore we need pass the actions object to this function as an argument.

Actions object can be gathered by passing a callback function to the component in a property named actionsRef.

For more information, see https://github.com/Financial-Times/x-dash/tree/master/components/x-interaction#triggering-actions-externally

Usage of this function is now documented in more detail in the readme file.