Fix live-blog-post padding
The bottom padding for the live blog post used to rely on sharing buttons to be visible. In the ft-app we don't display the share buttons in this component and the padding doesn't look right.

This fixes the bottom padding regardless of displaying or hiding the share buttons in that component.