Bubble event to parent w/ data-x-dash-id attribute
`x-interaction` listens for the `x-interaction.trigger-action` event on
the element with a `data-x-dash-id` attribute. Once it hears that event,
it kicks off an action based on the event type (eg. insertPost,
updatePost, deletePost). This wasn't working as expected because we
use the `data-x-live-blog-id` element to dispatch the
`x-interaction.trigger-action` event. For `x-interaction` to do it's
thing, we need to bubble the event to the `data-x-dash-id` element:

```
<div data-x-dash-id="live-blog-wrapper">
    <div class="x-live-blog-wrapper" data-live-blog-wrapper-id="live-blog-wrapper">
        {posts rendered here}
    </div>
</div>
```