Remove Snyk command from CircleCI publish step
A Snyk error is preventing me from releasing a new version of `x-dash`.
We thought this error was resolved but it looks not to be the case. Andy
has reported the issue to Snyk again and we are waiting on them.
Because we don't know if there will be a fix soon, I am going to remove
the Snyk command, publish a new release and then re-add the Snyk
command.

See more info here: https://financialtimes.slack.com/archives/C3TJ6KXEU/p1605111138223300