Remove failing Snyk CircleCI command
A Snyk error is preventing me from releasing a new version of `x-dash`.
Snyk is aware of the issue and Andy is working with them to debug.
Because we don't know if there will be a fix soon, I am going to remove
the Snyk command, publish a new release and then re-add the Snyk
command.