Give the x-live-blog-wrapper component a displayName
Without this the component wouldn't work with x-interaction when the code is transpiled for production.

It was working with development code I believe because it was relying in on teh standard Javascript name property of the function.

When the code was transpiled for production the functions were minified the function names were changed or made anonymous.

By hardcoding this displayName as a string it allows x-interaction to register the component and reference it by name.