Check for existing posts before inserting new ones.
We want to make sure that if a new post is received that has an id of one that already exists that we don't add it again, this will stop the user seeing multiples of the same post.

This makes this logic most resilient against issues upstream, for example the event stream sending multiple inserts, or inserts being sent due to race conditions.