fix: publish src folder for all packages
Since 8.0, consumers need to pull in the SCSS directly from the src
directory, rather than a bundled version of the generated CSS in the
dist directory. Unfortunately, some components (probably the ones
generated by the blueprint.js script,) have an .npmignore file that
tells npm not to publish the src directory, so these components weren't
having their SCSS published. Seeing as most components were fine without
an .npmignore file let's just delete it from all of the components so
that the SCSS can be published. Some other stuff might also be
spuriously included in the published bundle but it hasn't been an issue
with the other components!