Replace own myftLozengeTheme to n-myft-ui one
We haven't moved all myFT follow buttons to x-follow-button on ft.com yet. This causes we need to update stylings on both n-myft-ui and x-follow-button when some changes require to myftLozengeTheme. Actually some of the feature in n-myft-ui is not synced to x-follow-button. To avoid this issue, applying myftLozenge mixin from n-myft-ui.