fix: resolve webpack as peer dependency in x-engine
Instead of trying to manually resolve the version of webpack, we can
rely on the consuming application to have the correct reference to the
webpack dependency. We can also rely on the consumer to have the
dependencies defined for the x-engine config.

It looks like `resolvePeer` was made to make linking
`@financial-times/x-engine` possible. But now npm and workspaces can
handle this natively.