refactor: delete storybook webpack old monorepo override
We had a override in the storybook custom webpack config to handle
resolving packages when there are multiple node_modules directories per
monorepo package. Now we are using npm workspaces and it hoists most
packages to the root node_modules dependency, this removes the need for
the override and custom resolvers.