refactor: reintroduce resolvePeer for x-engine runtime and renderModule dependencies
Turns out when working with dynamic imports and using `npm link` in a
consuming project (like next-myft-page) to develop a x-dash component,
it is not able to resolve the runtime / renderModule as npm will try to
resolve the dependency within your local version of x-dash.

We could add peer dependencies for each runtime we support but the
previous method of working out the correct dependency path from the
consuming app and making it an alias worked just fine.