Modify close button in AS modal
The AS modal no longer uses o-overlay, so the style of the close button(which used o-overlay__close) had to be removed. Created a custom style class and copied the style of the close button from o-overlay so the style of the button does not break in the future when the o-overlay gets updated.